requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-google-genai
langchain-community
faiss-cpu
google-genai
python-multipart
requests
python-dotenv
